Enable allocation and write barrier hoisting out of loops
